requirements.txt
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
opencv-python==4.7.0.72
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
Pillow==9.5.0
wget==3.2
jupyter
notebook
tensorflow
keras
flask